Incorrect string format fixed (%d --> %lu).

+++ check_itunes_artworks/check_itunes_artworks.m
@@ -52,7 +52,7 @@ t_error check_selected_artworks(const t_prgm_options *options) {
 					[errorPrinter printErrorWithMessage:@"No artworks" track:ft];
 				
 				if (options->verbose)
-					[errorPrinter printErrorWithMessage:[NSString stringWithFormat:@"Being treated. Has %d artwork(s)", artworks.count] track:ft];
+					[errorPrinter printErrorWithMessage:[NSString stringWithFormat:@"Being treated. Has %lu artwork(s)", (unsigned long)artworks.count] track:ft];
 				
 				NSUInteger i = 0;
 				for (iTunesArtwork *curArtwork in artworks) {
@@ -60,15 +60,15 @@ t_error check_selected_artworks(const t_prgm_options *options) {
 					if (!options->check_all && i > 1) break;
 					
 					if (options->check_embed && curArtwork.downloaded)
-						[errorPrinter printErrorWithMessage:[NSString stringWithFormat:@"Artwork %d is not embedded in track", i] track:ft];
+						[errorPrinter printErrorWithMessage:[NSString stringWithFormat:@"Artwork %lu is not embedded in track", (unsigned long)i] track:ft];
 					
 					NSImage *curImage = [curArtwork data];
 					if (options->x_size > 0 && curImage.size.width != options->x_size)
-						[errorPrinter printErrorWithMessage:[NSString stringWithFormat:@"X Size of artwork %d is not correct (expected %lu, got %g)", i, options->x_size, curImage.size.width] track:ft];
+						[errorPrinter printErrorWithMessage:[NSString stringWithFormat:@"X Size of artwork %lu is not correct (expected %lu, got %g)", (unsigned long)i, options->x_size, curImage.size.width] track:ft];
 					if (options->y_size > 0 && curImage.size.height != options->y_size)
-						[errorPrinter printErrorWithMessage:[NSString stringWithFormat:@"Y Size of artwork %d is not correct (expected %lu, got %g)", i, options->y_size, curImage.size.height] track:ft];
+						[errorPrinter printErrorWithMessage:[NSString stringWithFormat:@"Y Size of artwork %lu is not correct (expected %lu, got %g)", (unsigned long)i, options->y_size, curImage.size.height] track:ft];
 					if (options->ratio >= 0 && !has_correct_ratio(curImage.size.width, curImage.size.height, options->ratio))
-						[errorPrinter printErrorWithMessage:[NSString stringWithFormat:@"Ratio of artwork %d is not correct (expected %g, got %g)", i, options->ratio, curImage.size.width / curImage.size.height] track:ft];
+						[errorPrinter printErrorWithMessage:[NSString stringWithFormat:@"Ratio of artwork %lu is not correct (expected %g, got %g)", (unsigned long)i, options->ratio, curImage.size.width / curImage.size.height] track:ft];
 				}
 			}
 		}
